Route to artifacts service; refactor pdf command
