+++ main.py
@@ -3,14 +3,9 @@ import logging
 import falcon.asgi
 import httpx
 import orjson
-import json
-import os
-import io
-import urllib.parse
-import boto3
-from botocore.config import Config  # type: ignore
-from jinja2 import Environment, FileSystemLoader
-from playwright.async_api import async_playwright
+
+from artifacts import v1_do_artifacts_connector
+from pdf import v1_do_pdf_connector
 
 # TODO: change this for prod
 logging.basicConfig(level=logging.INFO)
@@ -80,109 +75,6 @@ class v1_do_http_connector:
         }
 
 
-class v1_do_pdf_connector:
-    def __init__(self):
-        pass
-
-    async def on_post(self, req, resp):
-        params = await req.media
-        error = None
-        self.bucket = params.get("bucket")
-        self.object_name = params.get("object_name")
-        self.template_name = params.get("template_name")
-        self.config = json.loads(params.get("headers"))
-        self.test_data = params.get("test_data")
-
-        self.s3_client = boto3.client(
-            "s3",
-            endpoint_url=self.config.get("ENDPOINT_URL", None),
-            aws_access_key_id=self.config.get("AWS_ACCESS_KEY_ID"),
-            aws_secret_access_key=self.config.get("AWS_SECRET_ACCESS_KEY"),
-            config=Config(
-                region_name=self.config.get("AWS_DEFAULT_REGION"),
-                signature_version="s3v4",
-            ),
-        )
-
-        # Build the template
-        template_path = os.path.abspath("./templates")
-        env = Environment(loader=FileSystemLoader(template_path))
-        template = env.get_template(self.template_name)
-
-        # TODO: this can probably go away
-        # if not self.test_data:
-        #     render_data = task_data
-        # else:
-        #     render_data = self.test_data
-
-        render_data = self.test_data
-
-        # Render the document
-        rendered_document = template.render(render_data)
-        pdf_buffer = await self.html_to_pdf(rendered_document)
-
-        # take the buffer object and make a stream
-        pdf_stream = io.BytesIO(pdf_buffer)
-        # get the size so the client is happy
-        pdf_stream.seek(0, os.SEEK_END)
-        pdf_size = pdf_stream.tell()
-        # put stream back at start
-        pdf_stream.seek(0)
-        # Upload the file
-        try:
-            result = self.s3_client.put_object(
-                Bucket=self.bucket,
-                Key=self.object_name,
-                Body=pdf_stream,
-            )
-            # If no exception, upload succeeded. Now construct the response.
-            object_name = urllib.parse.quote_plus(self.object_name)
-
-            if "minio" in self.config.get("ENDPOINT_URL", None):
-                object_url = (
-                    f"http://localhost:9002/browser/{self.bucket}/{object_name}"
-                )
-            else:
-                object_url = f"https://s3-{self.config.get("AWS_DEFAULT_REGION")}.amazonaws.com/{self.bucket}/{self.object_name}"
-
-            response = json.dumps(
-                {
-                    "result": "success",
-                    "url": object_url,
-                    "bucket_name": self.bucket,
-                    "object_name": self.object_name,
-                    "etag": getattr(result, "etag", None),
-                    "version_id": getattr(result, "version_id", None),
-                }
-            )
-            status = "200"
-        except Exception as e:
-            response = "error"
-            error = json.dumps({"error": f"AWS Exception {e}"})
-            status = "500"
-
-        resp.media = {
-            "command_response": {
-                "body": response,
-                "mimetype": "application/json",
-                "http_status": status,
-            },
-            "command_response_version": 2,
-            "error": error,
-            "spiff__logs": [],
-        }
-
-    async def html_to_pdf(self, html_content):
-        """Create the PDF using Playwright library"""
-        async with async_playwright() as p:
-            browser = await p.chromium.launch()
-            page = await browser.new_page()
-            await page.set_content(html_content)
-            pdf_buffer = await page.pdf()
-            await browser.close()
-            return pdf_buffer
-
-
 #
 # App
 #
@@ -212,12 +104,30 @@ app.add_route("/v1/do/http/PatchRequest", v1_do_http_connector("PATCH"))
 app.add_route("/v1/do/http/PostRequest", v1_do_http_connector("POST"))
 app.add_route("/v1/do/http/PutRequest", v1_do_http_connector("PUT"))
 
+# Add new artifact routes
+artifacts = v1_do_artifacts_connector()
+app.add_route("/v1/do/artifacts/GenerateArtifact", artifacts, suffix="generate_artifact")
+app.add_route("/v1/do/artifacts/GetLinkToArtifact", artifacts, suffix="get_link")
+
+# Deprecated route - will be removed in future version
 app.add_route("/v1/do/pdf/pdf_to_s3", v1_do_pdf_connector())
 
 #
 # Static Data
 #
 
+generate_artifact_params = [
+    {"id": "id", "type": "str", "required": True},
+    {"id": "template", "type": "str", "required": True},
+    {"id": "data", "type": "dict", "required": True},
+    {"id": "generate_links", "type": "bool", "required": False},
+    {"id": "storage", "type": "str", "required": False},
+]
+
+get_link_params = [
+    {"id": "id", "type": "str", "required": True},
+    {"id": "storage", "type": "str", "required": False},
+]
 
 http_base_params = [
     {"id": "url", "type": "str", "required": True},
@@ -257,5 +167,8 @@ embedded_connectors = [
     {"id": "http/PatchRequest", "parameters": http_rw_params},
     {"id": "http/PostRequest", "parameters": http_rw_params},
     {"id": "http/PutRequest", "parameters": http_rw_params},
+    {"id": "artifacts/GenerateArtifact", "parameters": generate_artifact_params},
+    {"id": "artifacts/GetLinkToArtifact", "parameters": get_link_params},
+    # Deprecated - will be removed in future version
     {"id": "pdf/pdf_to_s3", "parameters": pdf_to_s3_params},
 ]
